feat(milk): resource-efficiency related changes, commit details for more info.
In this commit, we introduce pagination for the JSON API and a limiting of displayed users to 2,000 users for the Text API, in doing so, we performed additional changes to our queries and our schema, adding a `created_at` property to `Raid` and indexing `joined_at` for `RaidUser`.

Furthermore, we now grab `RaidUser` and the total count of the `Raid` at the same time.

Documentations have also been updated accordingly to accomodate these changes with an additional new documentation introduced to discuss the JSON API.